refactor(manifest): split workspace.rs into submodules
Split the 2,402-line workspace.rs into four cohesive submodules:

  - workspace/config.rs   — typed `WorkspaceConfig` + loading + caches
                            + `config_write_target` rule
  - workspace/edits.rs    — generic yaml/json edit helpers
                            (`edit_setting_map`, `edit_workspace_yaml`,
                            `upsert_map_entry`, `remove_map_entry`,
                            `workspace_yaml_submap`, …)
  - workspace/mutations.rs — domain-specific helpers for
                            `allowBuilds` and `patchedDependencies`
  - workspace/yaml_patch.rs — diff-and-apply on top of `yamlpatch`
                            (previously an inline `mod yaml_patch`)

workspace.rs becomes a thin coordinator: mod declarations, `pub use`
re-exports of the public API, and the existing test block at the
bottom. External callers continue to use `aube_manifest::workspace::X`
unchanged.

workspace.rs: 2,402 → 838 lines (most of the remainder is the test
module).

No behavior change; cargo fmt/clippy clean, full test suite passes
(99 manifest tests, including the yaml_patch detect-child-indent
suite that moved with its parent module).
