+++ crates/aube-manifest/src/workspace.rs
@@ -1,1604 +1,40 @@
-use crate::UpdateConfig;
-use serde::Deserialize;
-use std::collections::BTreeMap;
-use std::path::{Path, PathBuf};
-
-const WORKSPACE_YAML_NAMES: &[&str] = &["aube-workspace.yaml", "pnpm-workspace.yaml"];
-
-fn find_and_read(project_dir: &Path) -> Result<Option<(PathBuf, String)>, crate::Error> {
-    for name in WORKSPACE_YAML_NAMES {
-        let path = project_dir.join(name);
-        if path.exists() {
-            let content =
-                std::fs::read_to_string(&path).map_err(|e| crate::Error::Io(path.clone(), e))?;
-            return Ok(Some((path, content)));
-        }
-    }
-    Ok(None)
-}
-
-/// Extra privileges granted to one package pattern under `jailBuilds`.
-#[derive(Debug, Clone, Default, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct JailBuildPermission {
-    #[serde(default)]
-    pub env: Vec<String>,
-    #[serde(default)]
-    pub read: Vec<String>,
-    #[serde(default)]
-    pub write: Vec<String>,
-    #[serde(default)]
-    pub network: bool,
-}
-
-/// Configuration from `pnpm-workspace.yaml`.
-#[derive(Debug, Clone, Default, Deserialize)]
-#[serde(rename_all = "camelCase")]
-pub struct WorkspaceConfig {
-    /// Workspace package globs (e.g., `["packages/*"]`).
-    #[serde(default)]
-    pub packages: Vec<String>,
-
-    /// Include the root manifest in recursive/filter workspace operations.
-    #[serde(default)]
-    pub include_workspace_root: Option<bool>,
-
-    /// Default catalog for dependency version pinning.
-    #[serde(default)]
-    pub catalog: BTreeMap<String, String>,
-
-    /// Named catalogs for dependency version pinning.
-    #[serde(default)]
-    pub catalogs: BTreeMap<String, BTreeMap<String, String>>,
-
-    // -- Node-Modules Settings --
-    /// Linking strategy: "isolated" (default), "hoisted", or "pnp".
-    #[serde(default)]
-    pub node_linker: Option<String>,
-
-    /// Whether to use the global virtual store (default: false in pnpm, true in aube).
-    #[serde(default)]
-    pub enable_global_virtual_store: Option<bool>,
-
-    /// Package names whose presence in any importer forces
-    /// per-project materialization (disabling the global virtual
-    /// store for that install). Defaults to the common bundler /
-    /// framework direct-devDeps whose module resolvers follow
-    /// symlinks then walk up (Next.js's Turbopack, Vite, Rollup,
-    /// Webpack, Parcel, Nuxt, VitePress) — the global virtual store
-    /// makes `.aube/<pkg>` an absolute symlink that escapes the
-    /// project's filesystem root, which those resolvers can't walk
-    /// back from. Add more names as you discover tools with the same
-    /// restriction; set to `[]` to disable the heuristic. Declared
-    /// here so `settings.toml`'s workspaceYaml source stays in sync
-    /// with the actual deserialize surface.
-    #[serde(default)]
-    pub disable_global_virtual_store_for_packages: Option<Vec<String>>,
-
-    /// Package import method: "auto", "hardlink", "copy", "clone", "clone-or-copy".
-    #[serde(default)]
-    pub package_import_method: Option<String>,
-
-    /// Path to the virtual store directory (default: "node_modules/.aube").
-    #[serde(default)]
-    pub virtual_store_dir: Option<String>,
-
-    /// Top-level modules directory name (default: "node_modules").
-    /// aube accepts the setting for parity but only honors the
-    /// default value — see `settings.toml`'s `modulesDir` entry.
-    #[serde(default)]
-    pub modules_dir: Option<String>,
-
-    /// Whether to shamefully hoist all packages to root node_modules.
-    #[serde(default)]
-    pub shamefully_hoist: Option<bool>,
-
-    /// Master switch for the hidden modules directory at
-    /// `node_modules/.aube/node_modules/`. Default true.
-    #[serde(default)]
-    pub hoist: Option<bool>,
-
-    /// Patterns of packages to hoist.
-    #[serde(default)]
-    pub hoist_pattern: Option<Vec<String>>,
-
-    /// Whether workspace packages get symlinked into each importer's
-    /// `node_modules/`. Default true.
-    #[serde(default)]
-    pub hoist_workspace_packages: Option<bool>,
-
-    /// When true, the linker skips a workspace package's per-importer
-    /// `node_modules/<name>` symlink if the workspace root already
-    /// links the same package at the same resolved version. Default
-    /// false (pnpm parity).
-    #[serde(default)]
-    pub dedupe_direct_deps: Option<bool>,
-
-    /// When true, `aube deploy` copies every file in the source
-    /// workspace package into the target directory instead of
-    /// applying pack's `files` / `.npmignore` filter. Default false
-    /// (pnpm parity). Declared as a typed field so the settings-meta
-    /// parity test can see the workspace-yaml key.
-    #[serde(default)]
-    pub deploy_all_files: Option<bool>,
-
-    /// Patterns of packages to hoist to the root node_modules.
-    #[serde(default)]
-    pub public_hoist_pattern: Option<Vec<String>>,
-
-    // -- Store Settings --
-    /// Path to the content-addressable store.
-    #[serde(default)]
-    pub store_dir: Option<String>,
-
-    // -- Lockfile Settings --
-    /// Whether to use a lockfile (default: true).
-    #[serde(default)]
-    pub lockfile: Option<bool>,
-
-    /// Directory the lockfile is written to and read from. When unset
-    /// or equal to the project root, behaves as before. When set to a
-    /// different directory, the project becomes an importer keyed by
-    /// its relative path (mirrors pnpm's `lockfile-dir`).
-    #[serde(default)]
-    pub lockfile_dir: Option<String>,
-
-    /// Whether to prefer frozen lockfile (default: true).
-    #[serde(default)]
-    pub prefer_frozen_lockfile: Option<bool>,
-
-    /// Write a per-branch lockfile (`pnpm-lock.<branch>.yaml`) instead of
-    /// the default `pnpm-lock.yaml`. Reduces merge conflicts on long-lived
-    /// branches. Forward slashes in branch names are encoded as `!`.
-    #[serde(default)]
-    pub git_branch_lockfile: Option<bool>,
-
-    /// Branch-name glob list that triggers an automatic branch-lockfile
-    /// merge on matching branches. Companion to `gitBranchLockfile`.
-    /// See `settings.toml` for the pattern syntax (including `!`-prefix
-    /// negations). Declared as a typed field so the settings-meta parity
-    /// test can see the workspace-yaml key.
-    #[serde(default)]
-    pub merge_git_branch_lockfiles_branch_pattern: Option<Vec<String>>,
-
-    /// Write per-package lockfiles instead of one shared workspace lockfile.
-    /// Default `true` matches pnpm. The typed field is declared only so the
-    /// settings-meta parity test can see the workspace-yaml key — the
-    /// install path reads the value through `aube_settings::resolved`.
-    #[serde(default)]
-    pub shared_workspace_lockfile: Option<bool>,
-
-    /// Cap on lockfile peer-ID suffix byte length before the resolver
-    /// replaces the suffix with `_<sha256-hex>`. Default 1000 (pnpm
-    /// parity). Same typed/raw duality as `child_concurrency` — see
-    /// that field's comment.
-    #[serde(default)]
-    pub peers_suffix_max_length: Option<u64>,
-
-    // -- Dependency Resolution --
-    /// Override any dependency in the dependency graph.
-    #[serde(default)]
-    pub overrides: BTreeMap<String, String>,
-
-    /// `name@version` → patch-file-path map. pnpm v10 moved this out
-    /// of `package.json`'s `pnpm.patchedDependencies` so users can
-    /// document *why* a patch exists with YAML comments; aube merges
-    /// both locations, with workspace-yaml entries winning on key
-    /// conflict (same precedence as `overrides`).
-    #[serde(default, rename = "patchedDependencies")]
-    pub patched_dependencies: BTreeMap<String, String>,
-
-    /// os/cpu/libc widening set. pnpm v10 moved this alongside
-    /// `overrides` — users generating a cross-platform lockfile on
-    /// Linux CI want to widen in the workspace yaml (where the rest
-    /// of their shared config lives) rather than `package.json`.
-    /// Merged with `package.json`'s `pnpm.supportedArchitectures` /
-    /// `aube.supportedArchitectures` at install time.
-    #[serde(default, rename = "supportedArchitectures")]
-    pub supported_architectures: Option<SupportedArchitectures>,
-
-    /// Optional-dep names that should always be skipped, even when
-    /// their platform matches. Merged with `package.json`'s
-    /// `pnpm.ignoredOptionalDependencies` / `aube.*` at install time.
-    /// Distinct from `--no-optional`, which drops *all* optional deps.
-    #[serde(default, rename = "ignoredOptionalDependencies")]
-    pub ignored_optional_dependencies: Vec<String>,
-
-    /// Override for the pnpmfile path. pnpm lets users point at a
-    /// non-default location; aube's default is `cwd/.pnpmfile.mjs`
-    /// when present, otherwise `cwd/.pnpmfile.cjs`.
-    /// Relative paths resolve against the workspace root.
-    #[serde(default, rename = "pnpmfilePath")]
-    pub pnpmfile_path: Option<String>,
-
-    /// Extend package metadata during resolution.
-    #[serde(default)]
-    pub package_extensions: BTreeMap<String, yaml_serde::Value>,
-
-    /// Package deprecation ranges whose warnings should be muted.
-    #[serde(default)]
-    pub allowed_deprecated_versions: BTreeMap<String, String>,
-
-    /// Scope of install-time deprecation warnings: `none`, `direct`,
-    /// `all`, or `summary`. Declared as a typed field so the
-    /// settings-meta parity test sees the workspaceYaml key.
-    #[serde(default)]
-    pub deprecation_warnings: Option<String>,
-
-    /// Update-time policy knobs.
-    #[serde(default)]
-    pub update_config: Option<UpdateConfig>,
-
-    /// Trust-policy mode. Parsed for pnpm parity; resolver support is
-    /// limited to accepting the configured policy surface until registry
-    /// trust metadata is available.
-    #[serde(default)]
-    pub trust_policy: Option<String>,
-
-    /// Packages exempt from trust-policy checks.
-    #[serde(default)]
-    pub trust_policy_exclude: Vec<String>,
-
-    /// Ignore trust-policy checks for package versions older than this
-    /// many minutes.
-    #[serde(default)]
-    pub trust_policy_ignore_after: Option<u64>,
-
-    /// Reject transitive git/file/tarball dependency specs by default.
-    #[serde(default)]
-    pub block_exotic_subdeps: Option<bool>,
-
-    // -- Build Settings --
-    /// Whether to ignore all lifecycle scripts (default: false).
-    #[serde(default)]
-    pub ignore_scripts: Option<bool>,
-
-    // -- aube-specific knobs --
-    /// Skip the `aube run` / `aube exec` auto-install staleness check
-    /// at the workspace level. Same semantics as the `aubeNoAutoInstall`
-    /// setting resolved via `aube_settings::resolved`; both surfaces
-    /// round-trip through `WorkspaceConfig` to keep the
-    /// `workspace_yaml_keys_deserialize_onto_workspace_config` parity
-    /// test happy. See `AUBE_NO_AUTO_INSTALL` env-var alias.
-    #[serde(default)]
-    pub aube_no_auto_install: Option<bool>,
-
-    /// Bypass the project-level advisory lock on `node_modules/` for
-    /// every mutating aube command in this workspace. Same semantics as
-    /// the `aubeNoLock` setting resolved via `aube_settings::resolved`;
-    /// useful for CI matrices or deliberately-parallel test rigs
-    /// running from one shared workspace. See `AUBE_NO_LOCK` env-var
-    /// alias.
-    #[serde(default)]
-    pub aube_no_lock: Option<bool>,
-
-    /// Per-package allowlist for dependency lifecycle scripts. Keys are
-    /// pnpm-style patterns (`name`, `name@version`, `name@v1 || v2`);
-    /// values are `true` to allow or `false` to deny. Merged with
-    /// `package.json`'s `pnpm.allowBuilds` — workspace-level entries
-    /// take precedence for the same key.
-    #[serde(default)]
-    pub allow_builds: BTreeMap<String, yaml_serde::Value>,
-
-    /// pnpm's canonical allowlist format: a flat list of package names
-    /// whose lifecycle scripts are allowed to run. Merged with
-    /// `allow_builds` into the same `BuildPolicy`. Workspace-level
-    /// entries apply to every importer in the workspace.
-    #[serde(default, rename = "onlyBuiltDependencies")]
-    pub only_built_dependencies: Vec<String>,
-
-    /// pnpm's canonical denylist: lifecycle scripts from these packages
-    /// never run even if the allowlist includes them (explicit denies
-    /// always win in `BuildPolicy::decide`).
-    #[serde(default, rename = "neverBuiltDependencies")]
-    pub never_built_dependencies: Vec<String>,
-
-    /// Maximum number of dep lifecycle scripts running in parallel
-    /// during the post-link `allowBuilds` phase. Defaults to 5 when
-    /// unset. Mirrors pnpm's `childConcurrency` setting.
-    ///
-    /// The typed field isn't read directly by `install::run` —
-    /// every int/string setting in aube has two faces: this struct
-    /// field (for strict deserialization) and a
-    /// `aube_settings::resolved::<name>` accessor that reads from the
-    /// parallel raw YAML map in `ResolveCtx`. The duplication exists
-    /// so the `meta::workspace_yaml_keys_deserialize_onto_workspace_config`
-    /// test can catch settings.toml typos that would otherwise let a
-    /// YAML key fall through to `extra` and be silently ignored. Same
-    /// pattern as `minimum_release_age` / `auto_install_peers`.
-    #[serde(default, rename = "childConcurrency")]
-    pub child_concurrency: Option<u64>,
-
-    /// Cap concurrent tarball downloads. When unset, aube uses an
-    /// auto-scaled worker count x3 default, clamped to 16-64. Same
-    /// typed/raw duality as `child_concurrency`.
-    #[serde(default, rename = "networkConcurrency")]
-    pub network_concurrency: Option<u64>,
-
-    /// Cap package materialization/linking worker count. When unset,
-    /// aube uses platform-aware defaults in `aube-linker`. Same
-    /// typed/raw duality as `child_concurrency`.
-    #[serde(default, rename = "linkConcurrency")]
-    pub link_concurrency: Option<u64>,
-
-    /// Whether to verify each tarball's SHA-512 against the lockfile
-    /// integrity before importing into the store. Defaults to `true`
-    /// (pnpm parity); `false` skips the check.
-    #[serde(default, rename = "verifyStoreIntegrity")]
-    pub verify_store_integrity: Option<bool>,
-
-    /// Companion to `verifyStoreIntegrity`. When true, a missing
-    /// `dist.integrity` on an imported packument is a hard error
-    /// instead of a warning. Defaults to `false` for pnpm parity.
-    #[serde(default, rename = "strictStoreIntegrity")]
-    pub strict_store_integrity: Option<bool>,
-
-    /// Cache post-build side effects for dependency packages.
-    /// Accepted for pnpm-config parity but currently a no-op —
-    /// aube skips dep lifecycle scripts by default.
-    #[serde(default, rename = "sideEffectsCache")]
-    pub side_effects_cache: Option<bool>,
-
-    /// Run approved dependency lifecycle scripts in a restricted build
-    /// jail. Same typed/raw duality as `child_concurrency`.
-    #[serde(default, rename = "jailBuilds")]
-    pub jail_builds: Option<bool>,
-
-    /// Master switch that forces both `jailBuilds=true` and
-    /// `trustPolicy=no-downgrade`. Same typed/raw duality as
-    /// `child_concurrency`.
-    #[serde(default)]
-    pub paranoid: Option<bool>,
-
-    /// Dependency package patterns that should run outside the jail even
-    /// when `jailBuilds` is enabled. Same typed/raw duality as
-    /// `child_concurrency`.
-    #[serde(default, rename = "jailBuildExclusions")]
-    pub jail_build_exclusions: Vec<String>,
-
-    /// Extra env/path/network grants for packages that still run in the
-    /// jail. Keys use the same package-pattern syntax as `allowBuilds`.
-    #[serde(default, rename = "jailBuildPermissions")]
-    pub jail_build_permissions: BTreeMap<String, JailBuildPermission>,
-
-    // -- Catalog Settings --
-    /// Drop catalog entries that no importer references after resolve.
-    /// Wired through `aube_settings::resolved::cleanup_unused_catalogs`;
-    /// the typed field exists only so `meta::workspace_yaml_keys_...`
-    /// sees the key as a real field and doesn't fall through to `extra`.
-    #[serde(default)]
-    pub cleanup_unused_catalogs: Option<bool>,
-
-    // -- Workspace-protocol settings --
-    /// Resolve `aube add <name>` against local workspace siblings
-    /// before falling back to the registry. Wired through
-    /// `aube_settings::resolved::link_workspace_packages`; the typed
-    /// field exists so `meta::workspace_yaml_keys_...` sees the key
-    /// as a real field and doesn't fall through to `extra`.
-    #[serde(default)]
-    pub link_workspace_packages: Option<bool>,
-
-    /// Spec form written to `package.json` when `aube add` matches a
-    /// workspace sibling. The yaml value can be the booleans `true` /
-    /// `false` or the string `"rolling"`, so the typed field lands at
-    /// `yaml_serde::Value` and the resolver normalizes via
-    /// `aube_settings::resolved::SaveWorkspaceProtocol::from_str_normalized`.
-    #[serde(default)]
-    pub save_workspace_protocol: Option<yaml_serde::Value>,
-
-    // -- Peer Dependency Settings --
-    /// Whether to auto-install peer dependencies (default: true).
-    #[serde(default)]
-    pub auto_install_peers: Option<bool>,
-
-    /// Fail the install if any required peer dependency is missing or
-    /// resolves outside its declared range. Default: false (warn only).
-    #[serde(default)]
-    pub strict_peer_dependencies: Option<bool>,
-
-    /// Omit `link:` dependencies from the lockfile's importer
-    /// dependency maps on write. Default: false.
-    #[serde(default)]
-    pub exclude_links_from_lockfile: Option<bool>,
-
-    /// Collapse peer-equivalent subtree variants into a single
-    /// canonical dep_path (cross-subtree intersection). Default: true.
-    #[serde(default)]
-    pub dedupe_peer_dependents: Option<bool>,
-
-    /// Emit peer suffixes as `(version)` instead of `(name@version)`
-    /// in the lockfile. Default: false.
-    #[serde(default)]
-    pub dedupe_peers: Option<bool>,
-
-    /// Consult the root workspace importer's direct deps for peer
-    /// resolution before falling back to a graph-wide scan.
-    /// Default: true.
-    #[serde(default)]
-    pub resolve_peers_from_workspace_root: Option<bool>,
-
-    /// Record the full registry tarball URL on every registry-sourced
-    /// package in the lockfile's `resolution.tarball:` field. Default:
-    /// false. Round-trips through the lockfile `settings:` header so
-    /// the value is preserved once set.
-    #[serde(default)]
-    pub lockfile_include_tarball_url: Option<bool>,
-
-    // -- Supply Chain Settings --
-    /// Minimum age in minutes that a package version must have before
-    /// it's eligible for resolution. pnpm v11 default is 1440 (1 day).
-    #[serde(default)]
-    pub minimum_release_age: Option<u64>,
-
-    /// Package names exempt from `minimum_release_age`.
-    #[serde(default)]
-    pub minimum_release_age_exclude: Option<Vec<String>>,
-
-    /// When true, fail the install if no version satisfies a range
-    /// without violating `minimum_release_age`. Default: false (fall back
-    /// to the lowest satisfying version that ignores the cutoff).
-    #[serde(default)]
-    pub minimum_release_age_strict: Option<bool>,
-
-    /// OSV `MAL-*` advisory check policy for `aube add`. One of
-    /// `"on"` (default, fail open on fetch error), `"required"` (fail
-    /// closed on fetch error), or `"off"`.
-    #[serde(default)]
-    pub advisory_check: Option<String>,
-
-    /// OSV `MAL-*` advisory check policy for plain reinstalls,
-    /// backed by a local mirror of OSV's npm dump. Independent of
-    /// `advisory_check`, which fires on `aube add`, `aube update`,
-    /// and other fresh-resolution paths via the live OSV API.
-    /// Values: `"on"` (refresh-on-stale, fail-open on refresh
-    /// error), `"required"` (fail-closed on refresh error), or
-    /// `"off"` (default).
-    #[serde(default)]
-    pub advisory_check_on_install: Option<String>,
-
-    /// Bloom-filter prefilter for OSV `MAL-*` advisories on
-    /// lockfile-driven installs. Downloads a sub-MB filter from
-    /// `endevco/osv-bloom` (regenerated every 10 minutes) and
-    /// probes the resolved graph against it. Only bloom hits
-    /// escalate to the live OSV API for confirmation, so this
-    /// stays cheap enough to enable per-install. Values: `"on"`
-    /// (fail-open on refresh/probe error), `"required"`
-    /// (fail-closed on refresh/probe error), or `"off"` (default).
-    /// Coexists with `advisory_check` and `advisory_check_on_install`.
-    #[serde(default)]
-    pub advisory_bloom_check: Option<String>,
-
-    /// Force the live-API OSV `MAL-*` check on every install
-    /// (including frozen reinstalls). Default `false`. Set to
-    /// `true` for hardened CI where every install must observe
-    /// the latest advisories regardless of whether the lockfile
-    /// changed. Trades per-install latency for freshness.
-    #[serde(default)]
-    pub advisory_check_every_install: Option<bool>,
-
-    /// Weekly-downloads floor for `aube add`. Below this, aube prompts
-    /// for confirmation (or fails non-interactively). 0 disables.
-    #[serde(default)]
-    pub low_download_threshold: Option<u64>,
-
-    /// Glob patterns exempted from the `low_download_threshold` gate.
-    /// Each entry is matched against the full registry name (e.g.
-    /// `@scope/foo`). Matching names skip the weekly-downloads lookup;
-    /// the OSV `MAL-*` check still runs.
-    #[serde(default)]
-    pub allowed_unpopular_packages: Option<Vec<String>>,
-
-    /// Bun-compatible security scanner module spec (npm package name
-    /// or path) loaded via a `node` bridge at install / add time.
-    /// Empty string disables the integration.
-    #[serde(default)]
-    pub security_scanner: Option<String>,
-
-    /// pnpm-style peer dependency escape hatches. Read by
-    /// `PeerDependencyRules::resolve` during install; the actual matching
-    /// logic lives in `aube`. We only need the container here so the
-    /// settings-meta parity test can see the top-level key as a real
-    /// field (not falling into `extra`). Leaves of the map are
-    /// deserialized lazily on demand.
-    #[serde(default)]
-    pub peer_dependency_rules: Option<yaml_serde::Value>,
-
-    /// Capture unknown fields for forward compatibility.
-    #[serde(flatten)]
-    pub extra: BTreeMap<String, yaml_serde::Value>,
-}
-
-/// `supportedArchitectures.{os,cpu,libc}` arrays from
-/// pnpm-workspace.yaml. Same three-axis shape pnpm uses; each entry
-/// can be a concrete token (`"linux"`) or the literal `"current"`,
-/// which the resolver expands to the host triple.
-#[derive(Debug, Clone, Default, Deserialize)]
-pub struct SupportedArchitectures {
-    #[serde(default)]
-    pub os: Vec<String>,
-    #[serde(default)]
-    pub cpu: Vec<String>,
-    #[serde(default)]
-    pub libc: Vec<String>,
-}
-
-impl SupportedArchitectures {
-    pub fn is_empty(&self) -> bool {
-        self.os.is_empty() && self.cpu.is_empty() && self.libc.is_empty()
-    }
-}
-
-impl WorkspaceConfig {
-    /// Convert the raw `allow_builds` map to the same shape used for
-    /// `package.json`'s `pnpm.allowBuilds`, so callers can merge both
-    /// sources uniformly.
-    pub fn allow_builds_raw(&self) -> BTreeMap<String, crate::AllowBuildRaw> {
-        self.allow_builds
-            .iter()
-            .map(|(k, v)| {
-                let raw = match v {
-                    yaml_serde::Value::Bool(b) => crate::AllowBuildRaw::Bool(*b),
-                    // Strings are stored verbatim. `yaml_serde::to_string`
-                    // would re-encode the value as YAML — quoting strings
-                    // that need quoting, adding a trailing newline — and
-                    // that wrapped form would defeat the read-side
-                    // equality check against the canonical review
-                    // placeholder, plus surface extra quotes in any
-                    // warning the user sees.
-                    yaml_serde::Value::String(s) => crate::AllowBuildRaw::Other(s.clone()),
-                    other => {
-                        // Render via YAML serialization so the user sees
-                        // the same text they wrote (`[a, b]`) rather than
-                        // yaml_serde's Debug form. Matches the JSON side
-                        // in `AllowBuildRaw::from_json`.
-                        let rendered = yaml_serde::to_string(other)
-                            .unwrap_or_default()
-                            .trim()
-                            .to_string();
-                        crate::AllowBuildRaw::Other(rendered)
-                    }
-                };
-                (k.clone(), raw)
-            })
-            .collect()
-    }
-
-    /// Load workspace config from `aube-workspace.yaml` (preferred) or
-    /// `pnpm-workspace.yaml` (pnpm compatibility) in the given directory.
-    /// Returns `Default` if neither file exists. If both exist, the aube
-    /// file wins and the pnpm file is ignored.
-    ///
-    /// Memoized per-process. `find_workspace_packages`, lockfile-dir
-    /// resolution, catalog cleanup, jail-builds, and write-target
-    /// picking all hit this 4-8× per command with the same cwd.
-    /// Matches the existing `RAW_CACHE` pattern for the raw map.
-    pub fn load(project_dir: &Path) -> Result<Self, crate::Error> {
-        if let Some(hit) = typed_cache_lookup(project_dir) {
-            return Ok(hit);
-        }
-        let value = Self::load_uncached(project_dir)?;
-        typed_cache_insert(project_dir, value.clone());
-        Ok(value)
-    }
-
-    fn load_uncached(project_dir: &Path) -> Result<Self, crate::Error> {
-        let Some((path, content)) = find_and_read(project_dir)? else {
-            return Ok(Self::default());
-        };
-        if content.trim().is_empty() {
-            return Ok(Self::default());
-        }
-        crate::parse_yaml(&path, content)
-    }
-}
-
-type TypedCacheMap = std::collections::HashMap<std::path::PathBuf, WorkspaceConfig>;
-static TYPED_CACHE: std::sync::OnceLock<std::sync::Mutex<TypedCacheMap>> =
-    std::sync::OnceLock::new();
-
-fn typed_cache_lookup(project_dir: &Path) -> Option<WorkspaceConfig> {
-    let cache = TYPED_CACHE.get_or_init(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-    cache.lock().ok()?.get(project_dir).cloned()
-}
-
-fn typed_cache_insert(project_dir: &Path, value: WorkspaceConfig) {
-    let cache = TYPED_CACHE.get_or_init(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-    if let Ok(mut map) = cache.lock() {
-        map.insert(project_dir.to_path_buf(), value);
-    }
-}
-
-/// Load the workspace yaml as a raw top-level key/value map, without
-/// coercing into `WorkspaceConfig`'s typed fields. Intended for
-/// metadata-driven setting resolution (see `aube-settings`), where
-/// the caller walks a list of aliases from
-/// `SettingMeta::workspace_yaml_keys` and pulls out whichever key is
-/// present.
-///
-/// Returns an empty map if no file exists — same semantics as `load`.
-/// File-precedence rules match `load`: `aube-workspace.yaml` wins
-/// over `pnpm-workspace.yaml`.
-// Process-wide memoization for the raw workspace-yaml map. Hot-path
-// callers (`with_settings_ctx`, `aube_lock_filename`, `take_project_lock`,
-// and the install-path `load_both` caller) all hit this with the same
-// cwd. Same pattern as `aube_lockfile::aube_lock_filename`. Both
-// `load_raw` and `load_both` populate + read this cache so a later
-// `load_raw` after `load_both` doesn't re-read the file.
-type RawCacheMap =
-    std::collections::HashMap<std::path::PathBuf, BTreeMap<String, yaml_serde::Value>>;
-static RAW_CACHE: std::sync::OnceLock<std::sync::Mutex<RawCacheMap>> = std::sync::OnceLock::new();
-
-fn raw_cache_lookup(project_dir: &Path) -> Option<BTreeMap<String, yaml_serde::Value>> {
-    let cache = RAW_CACHE.get_or_init(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-    cache.lock().ok()?.get(project_dir).cloned()
-}
-
-fn raw_cache_insert(project_dir: &Path, value: BTreeMap<String, yaml_serde::Value>) {
-    let cache = RAW_CACHE.get_or_init(|| std::sync::Mutex::new(std::collections::HashMap::new()));
-    if let Ok(mut map) = cache.lock() {
-        map.insert(project_dir.to_path_buf(), value);
-    }
-}
-
-pub fn load_raw(project_dir: &Path) -> Result<BTreeMap<String, yaml_serde::Value>, crate::Error> {
-    if let Some(hit) = raw_cache_lookup(project_dir) {
-        return Ok(hit);
-    }
-    let Some((path, content)) = find_and_read(project_dir)? else {
-        raw_cache_insert(project_dir, BTreeMap::new());
-        return Ok(BTreeMap::new());
-    };
-    if content.trim().is_empty() {
-        raw_cache_insert(project_dir, BTreeMap::new());
-        return Ok(BTreeMap::new());
-    }
-    let parsed: BTreeMap<String, yaml_serde::Value> = crate::parse_yaml(&path, content)?;
-    raw_cache_insert(project_dir, parsed.clone());
-    Ok(parsed)
-}
-
-/// Load the workspace yaml once and return both the typed
-/// `WorkspaceConfig` view and the raw `BTreeMap` view, parsed from
-/// the same file contents. Callers that need both (e.g. `install::run`,
-/// which wants typed `allow_builds_raw()` *and* the raw map for
-/// metadata-driven setting resolution) avoid the two-read hit this
-/// way. Errors propagate instead of being silently swallowed.
-#[allow(clippy::type_complexity)]
-pub fn load_both(
-    project_dir: &Path,
-) -> Result<(WorkspaceConfig, BTreeMap<String, yaml_serde::Value>), crate::Error> {
-    let Some((path, content)) = find_and_read(project_dir)? else {
-        raw_cache_insert(project_dir, BTreeMap::new());
-        return Ok((WorkspaceConfig::default(), BTreeMap::new()));
-    };
-    if content.trim().is_empty() {
-        raw_cache_insert(project_dir, BTreeMap::new());
-        return Ok((WorkspaceConfig::default(), BTreeMap::new()));
-    }
-    let value: yaml_serde::Value = crate::parse_yaml(&path, content.clone())?;
-    let typed: WorkspaceConfig = yaml_serde::from_value(value.clone())
-        .map_err(|e| crate::Error::parse_yaml_err(&path, content.clone(), &e))?;
-    let raw: BTreeMap<String, yaml_serde::Value> = yaml_serde::from_value(value)
-        .map_err(|e| crate::Error::parse_yaml_err(&path, content, &e))?;
-    raw_cache_insert(project_dir, raw.clone());
-    Ok((typed, raw))
-}
-
-/// Path to the existing workspace yaml in `project_dir`, if any.
-/// `aube-workspace.yaml` wins over `pnpm-workspace.yaml` so an
-/// aube-native project's preferences override a co-existing pnpm
-/// fallback. Returns `None` when neither file exists — read-or-skip
-/// callers (catalog cleanup, ancestor walks) treat that as "nothing
-/// to read or rewrite".
-pub fn workspace_yaml_existing(project_dir: &Path) -> Option<PathBuf> {
-    for name in WORKSPACE_YAML_NAMES {
-        let path = project_dir.join(name);
-        if path.exists() {
-            return Some(path);
-        }
-    }
-    None
-}
-
-/// Resolve which workspace-yaml path a writer should mutate in
-/// `project_dir`. Existing `aube-workspace.yaml` wins over
-/// `pnpm-workspace.yaml`; when neither exists, falls back to
-/// `aube-workspace.yaml` — aube's own filename, parallel to the
-/// `aube-lock.yaml` shape we use for the lockfile.
-///
-/// Background: aube reads both `aube-workspace.yaml` (preferred)
-/// and `pnpm-workspace.yaml` (fallback) for backward compatibility
-/// with pnpm-style repos that already ship the latter. The
-/// generated default flips to the aube-prefixed name so a fresh
-/// project's filesystem layout matches aube's overall naming
-/// (`aube-lock.yaml`, `aube-workspace.yaml`) rather than mixing
-/// vendor namespaces.
-///
-/// Most writers should go through [`config_write_target`] instead,
-/// which only resolves to the workspace yaml when one already exists
-/// on disk. This raw helper is for the rare caller that genuinely
-/// needs a workspace yaml path even on a fresh project (e.g. the
-/// node-gyp bootstrap dummy file).
-pub fn workspace_yaml_target(project_dir: &Path) -> PathBuf {
-    workspace_yaml_existing(project_dir).unwrap_or_else(|| project_dir.join("aube-workspace.yaml"))
-}
-
-/// Where the next mutation of a workspace-level setting should land.
-/// `pnpm.<key>` in `package.json` and the workspace yaml hold the same
-/// shape for almost every aube-mutated setting (`patchedDependencies`,
-/// `allowBuilds`, future settings) so there is one rule that applies
-/// to all of them — see [`config_write_target`].
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub enum ConfigWriteTarget {
-    /// Mutate `pnpm.<key>` in `package.json` via [`edit_setting_map`].
-    PackageJson,
-    /// Mutate the existing workspace yaml at this path via
-    /// [`edit_workspace_yaml`].
-    WorkspaceYaml(PathBuf),
-}
-
-/// Pick which file a workspace-level config write should mutate. Pure
-/// file-existence rule: when the workspace yaml exists, write there
-/// (the pnpm v10+ canonical home, where YAML comments can document
-/// each entry); otherwise write to `package.json`. We deliberately do
-/// not introspect contents — a project with a workspace yaml gets all
-/// its workspace-level config there even when prior entries lived in
-/// `package.json`.
-///
-/// Used by every aube command that mutates a setting which can live in
-/// either file (`aube patch-commit`, `aube patch-remove`,
-/// `aube approve-builds`, install-time auto-deny seeding, …).
-pub fn config_write_target(project_dir: &Path) -> ConfigWriteTarget {
-    match workspace_yaml_existing(project_dir) {
-        Some(path) => ConfigWriteTarget::WorkspaceYaml(path),
-        None => ConfigWriteTarget::PackageJson,
-    }
-}
-
-/// Drop `entry_key` from `pnpm.<key>` and `aube.<key>` in
-/// `package.json`. Returns `Ok(true)` when at least one namespace held
-/// it. Empty inner maps and empty namespaces are scrubbed too. The
-/// rewrite is skipped entirely when nothing structural changes —
-/// mirrors the no-op-skip guarantee of [`edit_workspace_yaml`].
-///
-/// Walking both namespaces matters because the read side merges them
-/// (`aube.*` wins on conflict), so an entry recorded in either
-/// location is live; a one-namespace removal would leave a stale
-/// duplicate behind.
-pub fn remove_setting_entry(cwd: &Path, key: &str, entry_key: &str) -> Result<bool, crate::Error> {
-    let path = cwd.join("package.json");
-    if !path.exists() {
-        return Ok(false);
-    }
-    let raw = std::fs::read_to_string(&path).map_err(|e| crate::Error::Io(path.clone(), e))?;
-    let mut value = crate::parse_json::<serde_json::Value>(&path, raw)?;
-    let obj = value.as_object_mut().ok_or_else(|| {
-        crate::Error::YamlParse(path.clone(), "package.json is not an object".to_string())
-    })?;
-    let before = obj.clone();
-
-    let mut existed = false;
-    for ns in ["pnpm", "aube"] {
-        let mut ns_empty = false;
-        if let Some(ns_obj) = obj.get_mut(ns).and_then(|v| v.as_object_mut()) {
-            if let Some(inner) = ns_obj.get_mut(key).and_then(|v| v.as_object_mut()) {
-                if inner.remove(entry_key).is_some() {
-                    existed = true;
-                }
-                if inner.is_empty() {
-                    ns_obj.remove(key);
-                }
-            }
-            ns_empty = ns_obj.is_empty();
-        }
-        if ns_empty {
-            obj.remove(ns);
-        }
-    }
-
-    if *obj == before {
-        return Ok(existed);
-    }
-
-    let mut out = serde_json::to_string_pretty(&value)
-        .map_err(|e| crate::Error::YamlParse(path.clone(), format!("failed to serialize: {e}")))?;
-    out.push('\n');
-    std::fs::write(&path, out).map_err(|e| crate::Error::Io(path, e))?;
-    Ok(existed)
-}
-
-/// Mutate a namespaced map setting (e.g. `patchedDependencies`,
-/// `allowBuilds`) inside `package.json` and write back.
-///
-/// The closure receives a **merged** view of `pnpm.<key>` and
-/// `aube.<key>`, with `aube.*` winning on key conflict — the same
-/// precedence the read side already uses. After the closure runs,
-/// the merged result is written to a single namespace and the other
-/// is cleared, so a future read sees exactly one source of truth and
-/// can never silently shadow a stale entry. This matters because
-/// pnpm-aware tools (and pnpm itself) can introduce a `pnpm` key into
-/// a manifest after aube has already populated `aube.<key>`; without
-/// the merge-and-collapse, a re-record would leave the new value in
-/// `pnpm.<key>` while the stale `aube.<key>` entry kept winning on
-/// read.
-///
-/// The chosen namespace follows [`config_write_target`]'s rule:
-/// `pnpm` if a `pnpm` namespace is already declared in the manifest,
-/// `aube` otherwise. Empty namespaces and inner maps are scrubbed,
-/// and the rewrite is skipped entirely when nothing structural
-/// changes — mirrors the no-op-skip guarantee of [`edit_workspace_yaml`].
-pub fn edit_setting_map<F>(cwd: &Path, key: &str, f: F) -> Result<(), crate::Error>
-where
-    F: FnOnce(&mut serde_json::Map<String, serde_json::Value>),
-{
-    let path = cwd.join("package.json");
-    let raw = std::fs::read_to_string(&path).map_err(|e| crate::Error::Io(path.clone(), e))?;
-    let mut value = crate::parse_json::<serde_json::Value>(&path, raw)?;
-
-    let obj = value.as_object_mut().ok_or_else(|| {
-        crate::Error::YamlParse(path.clone(), "package.json is not an object".to_string())
-    })?;
-    let before = obj.clone();
-
-    // Build the merged view (pnpm first, aube overrides on conflict)
-    // before mutating, so the closure sees the same map the install
-    // path would.
-    let mut merged: serde_json::Map<String, serde_json::Value> = serde_json::Map::new();
-    for ns in ["pnpm", "aube"] {
-        if let Some(inner) = obj
-            .get(ns)
-            .and_then(serde_json::Value::as_object)
-            .and_then(|m| m.get(key))
-            .and_then(serde_json::Value::as_object)
-        {
-            for (k, v) in inner {
-                merged.insert(k.clone(), v.clone());
-            }
-        }
-    }
-
-    f(&mut merged);
-
-    let chosen_ns = if obj.contains_key("pnpm") {
-        "pnpm"
-    } else {
-        "aube"
-    };
-    let other_ns = if chosen_ns == "pnpm" { "aube" } else { "pnpm" };
-
-    // Drop `<key>` from the other namespace so the post-write state
-    // has one source of truth.
-    let mut other_ns_empty_after = false;
-    if let Some(other_obj) = obj.get_mut(other_ns).and_then(|v| v.as_object_mut()) {
-        other_obj.remove(key);
-        other_ns_empty_after = other_obj.is_empty();
-    }
-    if other_ns_empty_after {
-        obj.remove(other_ns);
-    }
-
-    // Write merged into the chosen namespace, or scrub it if empty.
-    if merged.is_empty() {
-        let mut chosen_ns_empty_after = false;
-        if let Some(chosen_obj) = obj.get_mut(chosen_ns).and_then(|v| v.as_object_mut()) {
-            chosen_obj.remove(key);
-            chosen_ns_empty_after = chosen_obj.is_empty();
-        }
-        if chosen_ns_empty_after {
-            obj.remove(chosen_ns);
-        }
-    } else {
-        let chosen_value = obj
-            .entry(chosen_ns.to_string())
-            .or_insert_with(|| serde_json::Value::Object(serde_json::Map::new()));
-        let chosen_obj = chosen_value.as_object_mut().ok_or_else(|| {
-            crate::Error::YamlParse(path.clone(), format!("`{chosen_ns}` is not an object"))
-        })?;
-        chosen_obj.insert(key.to_string(), serde_json::Value::Object(merged));
-    }
-
-    if *obj == before {
-        return Ok(());
-    }
-
-    let mut out = serde_json::to_string_pretty(&value)
-        .map_err(|e| crate::Error::YamlParse(path.clone(), format!("failed to serialize: {e}")))?;
-    out.push('\n');
-    std::fs::write(&path, out).map_err(|e| crate::Error::Io(path, e))?;
-    Ok(())
-}
-
-/// Force-write `names` in the project's `allowBuilds` map. Routes
-/// through [`config_write_target`]: workspace yaml when one exists,
-/// otherwise `package.json#pnpm.allowBuilds`. Returns the file that
-/// was written. Used by `aube approve-builds` and the
-/// `--allow-build=<pkg>` / `--deny-build=<pkg>` CLI flags — entries
-/// are forcibly set, overwriting any prior value.
-pub fn set_allow_builds(
-    project_dir: &Path,
-    names: &[String],
-    allow: bool,
-) -> Result<PathBuf, crate::Error> {
-    match config_write_target(project_dir) {
-        ConfigWriteTarget::WorkspaceYaml(path) => write_allow_builds_yaml(&path, names, allow),
-        ConfigWriteTarget::PackageJson => {
-            edit_setting_map(project_dir, "allowBuilds", |map| {
-                for name in names {
-                    map.insert(name.clone(), serde_json::Value::Bool(allow));
-                }
-            })?;
-            Ok(project_dir.join("package.json"))
-        }
-    }
-}
-
-/// Force-approve `names` in the project's `allowBuilds` map.
-pub fn add_to_allow_builds(project_dir: &Path, names: &[String]) -> Result<PathBuf, crate::Error> {
-    set_allow_builds(project_dir, names, true)
-}
-
-/// Upsert a single `<map>.<entry>` pair into the project's
-/// workspace-level config. Routes through [`config_write_target`]:
-/// workspace yaml when one exists, otherwise `<pnpm|aube>.<map>` in
-/// `package.json`. Returns the file that was written.
-///
-/// Used by `aube config set --local <map>.<entry> <value>` for any
-/// object-typed aube setting (`allowBuilds`, `overrides`,
-/// `packageExtensions`, …) so the dotted-key CLI syntax can write
-/// directly into the same maps `aube approve-builds` /
-/// install-time auto-deny seeding mutate. The value is passed in
-/// both yaml and json forms so the caller can choose the right scalar
-/// shape (bool vs string vs int) without this helper having to guess.
-pub fn upsert_map_entry(
-    project_dir: &Path,
-    map_name: &str,
-    entry_key: &str,
-    yaml_value: yaml_serde::Value,
-    json_value: serde_json::Value,
-) -> Result<PathBuf, crate::Error> {
-    match config_write_target(project_dir) {
-        ConfigWriteTarget::WorkspaceYaml(path) => {
-            edit_workspace_yaml(&path, |map| {
-                let submap = workspace_yaml_submap(map, map_name, &path)?;
-                submap.insert(yaml_serde::Value::String(entry_key.to_string()), yaml_value);
-                Ok(())
-            })?;
-            Ok(path)
-        }
-        ConfigWriteTarget::PackageJson => {
-            edit_setting_map(project_dir, map_name, |map| {
-                map.insert(entry_key.to_string(), json_value);
-            })?;
-            Ok(project_dir.join("package.json"))
-        }
-    }
-}
-
-/// Remove a single `<map>.<entry>` pair from the project's
-/// workspace-level config. Mirrors [`upsert_map_entry`]: sweeps both
-/// the workspace yaml (when one exists) and
-/// `<pnpm|aube>.<map>.<entry>` in `package.json` so a value set
-/// through either file can be deleted regardless of which one the
-/// current layout would have written to. Drops empty `<map>:`
-/// containers behind it so a removal doesn't leave a `{}` stub.
-///
-/// Returns `true` when at least one location held the entry. Used by
-/// `aube config delete --local <map>.<entry>` so dotted writes have
-/// a symmetric round-trip.
-pub fn remove_map_entry(
-    project_dir: &Path,
-    map_name: &str,
-    entry_key: &str,
-) -> Result<bool, crate::Error> {
-    let mut existed = false;
-    if let Some(yaml_path) = workspace_yaml_existing(project_dir) {
-        edit_workspace_yaml(&yaml_path, |map| {
-            let yaml_key = yaml_serde::Value::String(map_name.to_string());
-            let Some(submap) = map.get_mut(&yaml_key).and_then(|v| v.as_mapping_mut()) else {
-                return Ok(());
-            };
-            if submap.shift_remove(entry_key).is_some() {
-                existed = true;
-            }
-            if submap.is_empty() {
-                map.shift_remove(&yaml_key);
-            }
-            Ok(())
-        })?;
-    }
-    if remove_setting_entry(project_dir, map_name, entry_key)? {
-        existed = true;
-    }
-    Ok(existed)
-}
-
-/// Canonical placeholder string pnpm writes for unreviewed `allowBuilds`
-/// entries. Aube never writes it (we leave the manifest alone and rely
-/// on the warning + `aube approve-builds` flow instead), but pnpm-managed
-/// projects swapping to aube can carry these strings in their existing
-/// configs. The read-side in `aube-scripts::policy` recognizes this exact
-/// value and treats it as "skip without warning" rather than emitting
-/// an `UnsupportedValue` warning for every install.
-pub const ALLOW_BUILDS_REVIEW_PLACEHOLDER: &str = "set this to true or false";
-
-/// Insert or replace a single `patchedDependencies` entry in the
-/// workspace yaml at `path`. Creates the file (and the
-/// `patchedDependencies` mapping) if needed. The shared
-/// [`edit_workspace_yaml`] helper skips the rewrite when the closure
-/// produces no structural change, so an idempotent re-record after
-/// editing the patch file leaves yaml comments intact.
-pub fn upsert_workspace_patched_dependency(
-    path: &Path,
-    key: &str,
-    rel_patch_path: &str,
-) -> Result<PathBuf, crate::Error> {
-    edit_workspace_yaml(path, |map| {
-        let pd_map = workspace_yaml_submap(map, "patchedDependencies", path)?;
-        pd_map.insert(
-            yaml_serde::Value::String(key.to_string()),
-            yaml_serde::Value::String(rel_patch_path.to_string()),
-        );
-        Ok(())
-    })
-}
-
-/// Drop a `patchedDependencies` entry from the workspace yaml at
-/// `path`. Returns `Ok(true)` when the entry was removed (and the
-/// file was rewritten). When the removal empties
-/// `patchedDependencies` we drop the key from the document so we
-/// don't leave a `patchedDependencies: {}` stub behind.
-pub fn remove_workspace_patched_dependency(path: &Path, key: &str) -> Result<bool, crate::Error> {
-    let mut existed = false;
-    edit_workspace_yaml(path, |map| {
-        let pd_map = workspace_yaml_submap(map, "patchedDependencies", path)?;
-        existed = pd_map.shift_remove(key).is_some();
-        if pd_map.is_empty() {
-            map.shift_remove("patchedDependencies");
-        }
-        Ok(())
-    })?;
-    Ok(existed)
-}
-
-/// Get the inner mapping for a top-level workspace-yaml key, creating
-/// it if absent. Errors when the key exists but isn't a mapping (a
-/// hand-edited file shape we shouldn't silently replace).
-fn workspace_yaml_submap<'a>(
-    map: &'a mut yaml_serde::Mapping,
-    key: &str,
-    path: &Path,
-) -> Result<&'a mut yaml_serde::Mapping, crate::Error> {
-    let entry = map
-        .entry(yaml_serde::Value::String(key.to_string()))
-        .or_insert_with(|| yaml_serde::Value::Mapping(yaml_serde::Mapping::new()));
-    entry.as_mapping_mut().ok_or_else(|| {
-        crate::Error::YamlParse(path.to_path_buf(), format!("`{key}` must be a mapping"))
-    })
-}
-
-/// Apply `f` to the parsed top-level mapping of the workspace yaml at
-/// `path` and write it back. The helper exists so every workspace-yaml
-/// writer (allowBuilds, patchedDependencies, catalog cleanup, future
-/// settings) shares one comment-preserving rule: **user-authored
-/// comments and formatting in the file survive every edit**.
-///
-/// The closure mutates a parsed `yaml_serde::Mapping`. After it runs,
-/// the helper diffs before-vs-after and reduces the change set to a
-/// minimal sequence of `yamlpatch` operations applied directly to the
-/// original source. yamlpatch is comment- and format-preserving, so
-/// keys, comments, and whitespace that the closure didn't touch land
-/// back on disk byte-identical. A no-op closure produces an empty
-/// patch list and the file isn't rewritten at all.
-///
-/// For brand-new or empty files there is no source to preserve, so the
-/// helper falls back to `yaml_serde::to_string` for the initial write.
-pub fn edit_workspace_yaml<F>(path: &Path, f: F) -> Result<PathBuf, crate::Error>
-where
-    F: FnOnce(&mut yaml_serde::Mapping) -> Result<(), crate::Error>,
-{
-    use yaml_serde::{Mapping, Value};
-
-    let original_source: Option<String> = if path.exists() {
-        let content =
-            std::fs::read_to_string(path).map_err(|e| crate::Error::Io(path.to_path_buf(), e))?;
-        if content.trim().is_empty() {
-            None
-        } else {
-            Some(content)
-        }
-    } else {
-        None
-    };
-
-    let mut doc: Value = match original_source.as_deref() {
-        Some(content) => crate::parse_yaml(path, content.to_string())?,
-        None => Value::Mapping(Mapping::new()),
-    };
-
-    let map = doc.as_mapping_mut().ok_or_else(|| {
-        crate::Error::YamlParse(
-            path.to_path_buf(),
-            "top-level yaml must be a mapping".to_string(),
-        )
-    })?;
-
-    let before = map.clone();
-    f(map)?;
-    if *map == before {
-        return Ok(path.to_path_buf());
-    }
-
-    let after = std::mem::take(map);
-    write_workspace_yaml(path, original_source.as_deref(), &before, &after)?;
-    Ok(path.to_path_buf())
-}
-
-fn write_allow_builds_yaml(
-    path: &Path,
-    names: &[String],
-    allow: bool,
-) -> Result<PathBuf, crate::Error> {
-    edit_workspace_yaml(path, |map| {
-        let allow_builds = workspace_yaml_submap(map, "allowBuilds", path)?;
-        for name in names {
-            let key = yaml_serde::Value::String(name.clone());
-            allow_builds.insert(key, yaml_serde::Value::Bool(allow));
-        }
-        Ok(())
-    })
-}
-
-/// Persist a structural change against `path`. When `original_source`
-/// is `Some`, the change is encoded as a list of `yamlpatch`
-/// operations applied to the original text — comments and formatting
-/// the closure didn't touch survive the round trip. When it is `None`
-/// (fresh file or one that was empty), the after-state is serialized
-/// directly via `yaml_serde::to_string`; there is no source to
-/// preserve. Both paths atomic-write the result.
-fn write_workspace_yaml(
-    path: &Path,
-    original_source: Option<&str>,
-    before: &yaml_serde::Mapping,
-    after: &yaml_serde::Mapping,
-) -> Result<(), crate::Error> {
-    let bytes: Vec<u8> = match original_source {
-        Some(source) => yaml_patch::apply_diff(path, source, before, after)?,
-        None => {
-            let raw = yaml_serde::to_string(&yaml_serde::Value::Mapping(after.clone()))
-                .map_err(|e| crate::Error::YamlParse(path.to_path_buf(), e.to_string()))?;
-            indent_block_sequences(&raw).into_bytes()
-        }
-    };
-    aube_util::fs_atomic::atomic_write(path, &bytes)
-        .map_err(|e| crate::Error::Io(path.to_path_buf(), e))?;
-    Ok(())
-}
-
-/// Bump every block-sequence item line (`- ...`) by two spaces. Leaves
-/// already-indented lines and non-sequence lines alone. yaml_serde's
-/// output uses a single indent step per nesting level, so this produces
-/// the `parent:\n  - item` shape humans expect. Only used on the
-/// fresh-file write path; yamlpatch preserves the user's existing
-/// indentation otherwise.
-fn indent_block_sequences(input: &str) -> String {
-    let mut out = String::with_capacity(input.len() + 16);
-    for line in input.split_inclusive('\n') {
-        let stripped = line.trim_start_matches(' ');
-        if stripped.starts_with("- ") || stripped == "-\n" || stripped == "-" {
-            out.push_str("  ");
-        }
-        out.push_str(line);
-    }
-    out
-}
-
-/// Diff a parsed-then-mutated workspace yaml mapping into a minimal
-/// set of edits and apply them to the original source. Comments and
-/// formatting on untouched keys survive every edit.
-///
-/// The module is a thin wrapper around `yamlpatch` plus a manual
-/// block-mapping injector. yamlpatch handles `Remove` / `Replace` /
-/// scalar `Add` correctly. Its `Op::Add` for non-empty *mapping*
-/// values is broken upstream (it strips the nested indentation
-/// hierarchy and produces invalid YAML where a child key lands at the
-/// parent's column), so any new sub-mapping is rendered to a
-/// block-style YAML string here and inserted at the right byte
-/// offset instead.
-///
-/// The only public entry is [`apply_diff`].
-mod yaml_patch {
-    use std::path::Path;
-    use yaml_serde::{Mapping, Value};
-    use yamlpatch::{Op, Patch, apply_yaml_patches};
-    use yamlpath::{Component, Document, Route};
-
-    /// Indentation step new entries are rendered with. Two spaces
-    /// matches pnpm's canonical workspace yaml layout. Reading the
-    /// step from the source (so an existing four-space file stays
-    /// four-space) is left for a later pass — every aube install
-    /// plus existing pnpm workspaces use two.
-    const INDENT_STEP: usize = 2;
-
-    /// One unit of a structural diff. `Yp` operations go through
-    /// yamlpatch; `Add` operations are injected directly because
-    /// yamlpatch's `Op::Add` mishandles non-empty nested mappings.
-    enum Edit {
-        Yp(Patch<'static>),
-        Add {
-            route_keys: Vec<String>,
-            key: String,
-            value: serde_yaml::Value,
-        },
-    }
-
-    /// Compute the minimal edit list that turns `before` into `after`
-    /// and apply it to `source`. Returns the source unchanged when
-    /// the diff is empty.
-    pub(super) fn apply_diff(
-        path: &Path,
-        source: &str,
-        before: &Mapping,
-        after: &Mapping,
-    ) -> Result<Vec<u8>, crate::Error> {
-        let mut edits = Vec::new();
-        diff_into(before, after, &[], path, &mut edits)?;
-        if edits.is_empty() {
-            return Ok(source.as_bytes().to_vec());
-        }
-
-        // Step 1: yamlpatch-handled ops (Remove + Replace + scalar
-        // Add). These are surgical and order-independent: yamlpatch
-        // applies them sequentially against the tree-sitter doc,
-        // re-querying after each step.
-        let yp_patches: Vec<Patch<'static>> = edits
-            .iter()
-            .filter_map(|e| match e {
-                Edit::Yp(p) => Some(p.clone()),
-                _ => None,
-            })
-            .collect();
-        let mut current = if yp_patches.is_empty() {
-            source.to_string()
-        } else {
-            let document =
-                Document::new(source.to_string()).map_err(|e| yp_err(path, e.to_string()))?;
-            apply_yaml_patches(&document, &yp_patches)
-                .map_err(|e| yp_err(path, e.to_string()))?
-                .source()
-                .to_string()
-        };
-
-        // Step 2: direct injections for new keys whose value is a
-        // mapping. Sort outer-most first so a parent that only just
-        // came into existence is queryable for its children. Within
-        // the same depth, preserve insertion order.
-        let mut adds: Vec<(Vec<String>, String, serde_yaml::Value)> = edits
-            .into_iter()
-            .filter_map(|e| match e {
-                Edit::Add {
-                    route_keys,
-                    key,
-                    value,
-                } => Some((route_keys, key, value)),
-                _ => None,
-            })
-            .collect();
-        adds.sort_by_key(|(r, _, _)| r.len());
-        for (route_keys, key, value) in adds {
-            current = inject_entry(&current, &route_keys, &key, &value, path)?;
-        }
-
-        Ok(current.into_bytes())
-    }
-
-    /// Walk `before` and `after` recursively, pushing `Edit`s for
-    /// every structural difference. Mapping-valued additions become
-    /// `Edit::Add` (handled outside yamlpatch); everything else maps
-    /// to a yamlpatch `Patch`. Non-string keys cause a hard error
-    /// rather than silent data loss.
-    fn diff_into(
-        before: &Mapping,
-        after: &Mapping,
-        route: &[String],
-        path: &Path,
-        out: &mut Vec<Edit>,
-    ) -> Result<(), crate::Error> {
-        let route_obj: Route<'static> = Route::from(
-            route
-                .iter()
-                .cloned()
-                .map(Component::from)
-                .collect::<Vec<_>>(),
-        );
-        for (k, _) in before.iter() {
-            let key = key_str(path, k)?;
-            if !after.contains_key(k) {
-                out.push(Edit::Yp(Patch {
-                    route: route_obj.with_key(key.to_string()),
-                    operation: Op::Remove,
-                }));
-            }
-        }
-        for (k, after_v) in after.iter() {
-            let key = key_str(path, k)?;
-            match before.get(k) {
-                None => out.push(Edit::Add {
-                    route_keys: route.to_vec(),
-                    key: key.to_string(),
-                    value: to_serde_value(path, after_v)?,
-                }),
-                Some(before_v) if before_v != after_v => {
-                    if let (Some(bm), Some(am)) = (before_v.as_mapping(), after_v.as_mapping()) {
-                        let mut sub = route.to_vec();
-                        sub.push(key.to_string());
-                        diff_into(bm, am, &sub, path, out)?;
-                    } else if matches!(after_v.as_mapping(), Some(m) if !m.is_empty()) {
-                        // Type-change to a non-empty sub-mapping (e.g.
-                        // scalar -> nested mapping). yamlpatch's
-                        // Op::Replace serializes the mapping value via
-                        // the same path as Op::Add, which strips nested
-                        // indentation and lands the children at the
-                        // parent's column. Split into Remove + manual
-                        // injection so step 2 can re-emit the children
-                        // with their canonical indent.
-                        out.push(Edit::Yp(Patch {
-                            route: route_obj.with_key(key.to_string()),
-                            operation: Op::Remove,
-                        }));
-                        out.push(Edit::Add {
-                            route_keys: route.to_vec(),
-                            key: key.to_string(),
-                            value: to_serde_value(path, after_v)?,
-                        });
-                    } else {
-                        out.push(Edit::Yp(Patch {
-                            route: route_obj.with_key(key.to_string()),
-                            operation: Op::Replace(after_v.clone()),
-                        }));
-                    }
-                }
-                _ => {}
-            }
-        }
-        Ok(())
-    }
-
-    /// Inject a fresh `<key>: <value>` block-style entry into
-    /// `source` at the end of the route's mapping. Top-level routes
-    /// (empty) append at end-of-file. Nested routes look up the
-    /// parent feature via yamlpath, then insert just past its end
-    /// span at the parent's child indent.
-    fn inject_entry(
-        source: &str,
-        route_keys: &[String],
-        key: &str,
-        value: &serde_yaml::Value,
-        path: &Path,
-    ) -> Result<String, crate::Error> {
-        if route_keys.is_empty() {
-            let entry = render_entry(key, value, 0);
-            let mut result = source.to_string();
-            if !result.is_empty() && !result.ends_with('\n') {
-                result.push('\n');
-            }
-            result.push_str(&entry);
-            return Ok(result);
-        }
-
-        let document =
-            Document::new(source.to_string()).map_err(|e| yp_err(path, e.to_string()))?;
-        let route_obj: Route<'static> = Route::from(
-            route_keys
-                .iter()
-                .cloned()
-                .map(Component::from)
-                .collect::<Vec<_>>(),
-        );
-        let feature = document
-            .query_exact(&route_obj)
-            .map_err(|e| yp_err(path, e.to_string()))?
-            .ok_or_else(|| {
-                yp_err(
-                    path,
-                    format!("parent route {route_keys:?} not found in source"),
-                )
-            })?;
-        // `extract_with_leading_whitespace` walks the byte span back
-        // over any pure-space prefix on the parent's first line, so
-        // the snapshot mirrors the original column the children sit
-        // at — `extract` alone would start mid-line and drop the
-        // indent the new entry needs to inherit.
-        let parent_content = document.extract_with_leading_whitespace(&feature);
-        let child_indent = detect_child_indent(parent_content, route_keys.len());
-        let entry = render_entry(key, value, child_indent);
-
-        let mut insert_at = feature.location.byte_span.1;
-        // Trim back over trailing whitespace so the new entry lands
-        // just after the parent block's last content line, before any
-        // trailing blank lines that belong to the document footer.
-        let bytes = source.as_bytes();
-        while insert_at > 0 && matches!(bytes[insert_at - 1], b'\n' | b' ') {
-            insert_at -= 1;
-        }
-        let mut result = source.to_string();
-        let mut prefix = String::new();
-        if insert_at == 0 || bytes[insert_at - 1] != b'\n' {
-            prefix.push('\n');
-        }
-        prefix.push_str(&entry);
-        result.insert_str(insert_at, &prefix);
-        Ok(result)
-    }
-
-    /// Render `<key>: <value>` as block-style YAML lines, each
-    /// indented by `indent` spaces. Non-empty mapping values nest
-    /// recursively; non-empty sequence values emit as block
-    /// sequences with `- ` items at child indent; everything else
-    /// is emitted as a scalar value after the colon.
-    fn render_entry(key: &str, value: &serde_yaml::Value, indent: usize) -> String {
-        let pad = " ".repeat(indent);
-        match value {
-            serde_yaml::Value::Mapping(m) if !m.is_empty() => {
-                let mut out = format!("{pad}{}:\n", scalar_key_str(key));
-                for (k, v) in m {
-                    let child_key = match k {
-                        serde_yaml::Value::String(s) => s.clone(),
-                        other => render_scalar(other),
-                    };
-                    out.push_str(&render_entry(&child_key, v, indent + INDENT_STEP));
-                }
-                out
-            }
-            serde_yaml::Value::Sequence(seq) if !seq.is_empty() => {
-                // Block-sequence under a mapping key. Without an
-                // explicit arm the catch-all below would feed the
-                // sequence through `render_scalar`, which emits a
-                // multi-line `- a\n- b` chunk that lands inline on the
-                // `key:` line and produces structurally invalid YAML.
-                let mut out = format!("{pad}{}:\n", scalar_key_str(key));
-                let item_pad = " ".repeat(indent + INDENT_STEP);
-                for item in seq {
-                    if matches!(
-                        item,
-                        serde_yaml::Value::Mapping(_) | serde_yaml::Value::Sequence(_)
-                    ) {
-                        // Nested mapping/sequence as a list item: defer
-                        // to serde_yaml for inner shape, then attach
-                        // the dash to the first emitted line and pad
-                        // every continuation line so it stays inside
-                        // the same item.
-                        let raw = serde_yaml::to_string(item).unwrap_or_default();
-                        let mut first = true;
-                        for line in raw.lines() {
-                            if first {
-                                first = false;
-                                out.push_str(&item_pad);
-                                out.push_str("- ");
-                            } else {
-                                out.push_str(&item_pad);
-                                out.push_str("  ");
-                            }
-                            out.push_str(line);
-                            out.push('\n');
-                        }
-                    } else {
-                        out.push_str(&item_pad);
-                        out.push_str("- ");
-                        out.push_str(&render_scalar(item));
-                        out.push('\n');
-                    }
-                }
-                out
-            }
-            _ => format!("{pad}{}: {}\n", scalar_key_str(key), render_scalar(value)),
-        }
-    }
-
-    /// Re-serialize a single scalar through serde_yaml so YAML
-    /// quoting (escapes, leading-special-char handling) matches what
-    /// the rest of the file already uses. Trailing newlines from the
-    /// emitter are stripped — the caller owns its own line break.
-    fn render_scalar(value: &serde_yaml::Value) -> String {
-        let raw = serde_yaml::to_string(value).unwrap_or_default();
-        raw.trim_end().to_string()
-    }
-
-    /// Render a mapping key for emission, quoting only when the YAML
-    /// 1.2 plain-scalar grammar requires it. Defers to serde_yaml's
-    /// emitter so the rules stay in lockstep with the rest of the
-    /// file: identifiers like `b@2.0.0` and `is-positive@3.1.0`
-    /// round-trip unquoted (the `@` is reserved only at the *start*
-    /// of a scalar), while keys that lead with a reserved indicator
-    /// or contain flow/quote/comment characters get the canonical
-    /// quoted form serde_yaml would have produced.
-    fn scalar_key_str(key: &str) -> String {
-        let raw = serde_yaml::to_string(&serde_yaml::Value::String(key.to_string()))
-            .unwrap_or_else(|_| format!("{key}\n"));
-        raw.trim_end().to_string()
-    }
-
-    /// Inspect a parent block-mapping's source text to decide what
-    /// indent its new children should land at. The slice handed in
-    /// here comes from `extract_with_leading_whitespace`, so its
-    /// first line is already a child of the parent route — return
-    /// that first non-empty/non-comment line's leading whitespace.
-    /// Falls back to `parent_depth * INDENT_STEP` (the parent's own
-    /// column plus one more step) when the parent is empty: a depth-2
-    /// parent with no children otherwise had its children land at
-    /// column 2 alongside the parent itself rather than column 4.
-    fn detect_child_indent(parent_content: &str, parent_depth: usize) -> usize {
-        for line in parent_content.lines() {
-            let trimmed = line.trim_start();
-            if trimmed.is_empty() || trimmed.starts_with('#') {
-                continue;
-            }
-            return line.len() - trimmed.len();
-        }
-        parent_depth * INDENT_STEP
-    }
-
-    /// Extract a string view of a mapping key, erroring out for any
-    /// non-string variant. yaml_serde mappings allow non-string keys
-    /// in principle but every workspace-yaml shape aube edits uses
-    /// string keys exclusively, and silently dropping anything else
-    /// would lose data on the rewrite.
-    fn key_str<'a>(path: &Path, value: &'a Value) -> Result<&'a str, crate::Error> {
-        match value {
-            Value::String(s) => Ok(s.as_str()),
-            other => Err(yp_err(
-                path,
-                format!("workspace yaml mapping key must be a string, got {other:?}"),
-            )),
-        }
-    }
-
-    /// Bridge `yaml_serde::Value` (our typed parse type) to
-    /// `serde_yaml::Value` (the manual injector's render type).
-    /// yaml_serde is
-    /// the maintained fork of serde_yaml 0.9, so a YAML round-trip is
-    /// lossless for every variant we use (scalars, sequences,
-    /// mappings, tagged values). Errors on either side propagate
-    /// instead of panicking — they're vanishingly rare but a
-    /// workspace edit is a poor place to crash the process.
-    fn to_serde_value(path: &Path, value: &Value) -> Result<serde_yaml::Value, crate::Error> {
-        let raw = yaml_serde::to_string(value).map_err(|e| yp_err(path, e.to_string()))?;
-        serde_yaml::from_str(&raw).map_err(|e| yp_err(path, e.to_string()))
-    }
-
-    fn yp_err(path: &Path, msg: String) -> crate::Error {
-        crate::Error::YamlParse(path.to_path_buf(), msg)
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[test]
-        fn detect_child_indent_reads_existing_child_indent() {
-            assert_eq!(detect_child_indent("    foo: 1\n", 2), 4);
-            assert_eq!(detect_child_indent("  foo: 1\n", 1), 2);
-        }
-
-        #[test]
-        fn detect_child_indent_skips_blank_and_comment_lines() {
-            assert_eq!(detect_child_indent("\n    # note\n    foo: 1\n", 2), 4);
-        }
-
-        #[test]
-        fn detect_child_indent_falls_back_to_parent_depth() {
-            // Depth-2 parent (e.g. `catalogs.evens`) with no children:
-            // children should land at column 4, not column 2.
-            assert_eq!(detect_child_indent("", 2), 4);
-            // Depth-1 parent: children at column 2.
-            assert_eq!(detect_child_indent("", 1), 2);
-            // Depth-3 parent: children at column 6.
-            assert_eq!(detect_child_indent("", 3), 6);
-        }
-    }
-}
+//! Workspace-level configuration (`pnpm-workspace.yaml` /
+//! `aube-workspace.yaml`) read + write.
+//!
+//! Split into four submodules:
+//! - [`config`] — typed `WorkspaceConfig`, parsing, in-memory caches,
+//!   and the `config_write_target` rule.
+//! - [`edits`] — generic yaml/json edit helpers shared by every
+//!   workspace-level mutation.
+//! - [`mutations`] — the domain-specific mutations: `allowBuilds`
+//!   and `patchedDependencies`.
+//! - [`yaml_patch`] — comment-preserving diff-and-apply on top of
+//!   `yamlpatch`, with a manual injector for new sub-mappings (which
+//!   `yamlpatch::Op::Add` mishandles).
+//!
+//! The submodules are private; the public API is re-exported from
+//! this module so callers continue to use `aube_manifest::workspace::X`.
+
+mod config;
+mod edits;
+mod mutations;
+mod yaml_patch;
+
+pub use config::{
+    ConfigWriteTarget, JailBuildPermission, SupportedArchitectures, WorkspaceConfig,
+    config_write_target, load_both, load_raw, workspace_yaml_existing, workspace_yaml_target,
+};
+pub use edits::{
+    edit_setting_map, edit_workspace_yaml, remove_map_entry, remove_setting_entry, upsert_map_entry,
+};
+pub use mutations::{
+    ALLOW_BUILDS_REVIEW_PLACEHOLDER, add_to_allow_builds, remove_workspace_patched_dependency,
+    set_allow_builds, upsert_workspace_patched_dependency,
+};
 
 #[cfg(test)]
 mod tests {
+    use super::edits::workspace_yaml_submap;
     use super::*;
 
     #[test]
